+++ CHANGES.md
@@ -1,11 +1,13 @@
 **Added**  
-* 
+* generic error reporting for ArmaAddonsIndexingCallback interface
 
 **Changed**  
 * 
 
 **Fixed**  
-* 
+* IndexArmaAddonsStatusDialog:
+    * made error and warning count display correctly at start
+* ArmaTools CfgConvert: wasn't converting the bin file to a de-binarized file 
 
 **Notes**
 * scope works, but it doesn't work for deep nested scopes